requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<3.16
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
PyMySQL>=1.1,<2.0
